services/transcriptions/lib/requirements.model.txt
piano_transcription_inference @ git+https://github.com/sebat2004/piano_transcription_inference
torch
numpy
librosa==0.9.2
boto3
google
note_seq
psycopg[binary,pool]==3.2.1
psycopg2-binary==2.9.10
python-dotenv